Final Push

+++ javascript/addRoleList.js
@@ -8,14 +8,14 @@ const getRole = async () => {
       for (let i = 0; i < Role.length; i++) {
         const tr = document.createElement('tr');
         const trContent = `
-                                  <td>${i + 1}</td>
-                                  <td>${Role[i].roleName}</td>
-                                  <td>
-                                  <a style="cursor: pointer;"><span delete="true" style="color: red;" id="${
-                                    Role[i]._id
-                                  }">Fshi</span></a>
-                                  </td>
-                                  `;
+          <td>${i + 1}</td>
+          <td>${Role[i].roleName}</td>
+          <td>
+            <a style="cursor: pointer;"><span delete="true" style="color: red;" id="${
+              Role[i]._id
+            }">Fshi</span></a>
+          </td>
+        `;
         tr.innerHTML = trContent;
         document.querySelector('table tbody').appendChild(tr);
       }
@@ -27,19 +27,37 @@ const getRole = async () => {
   const deleteButtons = document.querySelectorAll('[delete="true"]');
   deleteButtons.forEach((element) => {
     element.addEventListener('click', async (event) => {
-      const res = await fetch('/role/delete-role', {
-        method: 'DELETE',
-        body: JSON.stringify({
-          id: event.target.id,
-        }),
-        headers: {
-          'Content-Type': 'application/json',
-        },
+      const result = await Swal.fire({
+        title: 'Jeni të sigurt që dëshironi ta fshini?',
+        icon: 'warning',
+        iconColor: '#e72d18',
+        showCancelButton: true,
+        confirmButtonText: 'Fshini',
+        cancelButtonText: 'Anuloje',
       });
-      const data = await res.json();
 
-      if (data.status === 'success') {
-        getRole();
+      if (result.isConfirmed) {
+        const res = await fetch('/role/delete-role', {
+          method: 'DELETE',
+          body: JSON.stringify({
+            id: event.target.id,
+          }),
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        });
+        const data = await res.json();
+
+        if (data.status === 'success') {
+          getRole();
+          Swal.fire('Fshirë!', 'Profesioni është fshirë me sukses.', 'success');
+        } else {
+          Swal.fire(
+            'Gabim!',
+            'Ka ndodhur një gabim gjatë fshirjes së profesionit.',
+            'error'
+          );
+        }
       }
     });
   });
